move all extractors to MatchCaseCompletions
